+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Changes in 1.7.3
 
-- Add `MsBackendDF` backend class providing a more efficient in-memory data
+- Add `MsBackendMemory` backend class providing a more efficient in-memory data
   representation than `MsBackendDataFrame`.
 
 ## Changes in 1.7.2